Rename devsettings files (#47)
